@@ -16,26 +16,28 @@
     "@emotion/styled": "^11.10.0",
     "@mui/icons-material": "^5.8.4",
     "@mui/material": "^5.10.1",
+    "@next/env": "^14.0.1",
     "@nimiq/style": "^0.8.5",
     "@react-three/drei": "^9.80.3",
     "@react-three/fiber": "^8.13.6",
     "@starknet-react/chains": "^0.1.0-next.0",
-    "@starknet-react/core": "^2.0.0-next.1",
+    "@starknet-react/core": "^2.0.0-next.6",
     "@use-gesture/react": "^10.2.27",
     "@vercel/analytics": "^0.1.5",
     "big.js": "^6.2.1",
     "bn.js": "^5.2.1",
     "chart.js": "^4.3.0",
+    "eslint-config-next": "^14.0.1",
     "get-starknet-core": "^3.2.0",
     "jotai": "^2.5.0",
     "lottie-react": "^2.4.0",
     "maath": "^0.7.0",
     "mongodb": "^4.12.1",
-    "next": "12.2.5",
+    "next": "^14.0.1",
     "nextjs-cors": "^2.1.2",
-    "react": "18.2.0",
+    "react": "^18.2.0",
     "react-chartjs-2": "^5.2.0",
-    "react-dom": "18.2.0",
+    "react-dom": "^18.2.0",
     "react-icons": "^4.4.0",
     "react-intersection-observer": "^9.5.2",
     "react-loader-spinner": "^5.2.0",
@@ -51,8 +53,8 @@
     "@types/big.js": "^6.1.6",
     "@types/bn.js": "^5.1.0",
     "@types/node": "18.7.8",
-    "@types/react": "18.0.17",
-    "@types/react-dom": "18.0.6",
+    "@types/react": "^18.2.37",
+    "@types/react-dom": "^18.2.15",
     "@types/three": "^0.155.0",
     "@typescript-eslint/eslint-plugin": "^5.47.0",
     "@typescript-eslint/parser": "^5.47.0",
