@@ -1,16 +1,17 @@
-import React, { useContext } from "react";
-import type { NextPage } from "next";
-import styles from "../styles/Home.module.css";
-import FeaturedQuest from "../components/quests/featuredQuest";
+"use client";
 
-import { useRouter } from "next/router";
+import React, { useContext } from "react";
+import styles from "../styles/Home.module.css";
+
+import { useRouter } from "next/navigation";
 import HowToParticipate from "../components/pages/home/howToParticipate";
 import QuestCategories from "../components/pages/home/questCategories";
 import TrendingQuests from "../components/pages/home/trending";
 import Blur from "../components/shapes/blur";
 import { QuestsContext } from "../context/QuestsProvider";
+import FeaturedQuest from "../components/UI/featured_banner/featuredQuest";
 
-const Quests: NextPage = () => {
+export default function Page() {
   const router = useRouter();
   const { featuredQuest, categories, trendingQuests } =
     useContext(QuestsContext);
@@ -21,19 +22,23 @@ const Quests: NextPage = () => {
         <div className={styles.blur1}>
           <Blur />
         </div>
-        <FeaturedQuest
-          key={featuredQuest?.id}
-          title={featuredQuest?.title_card}
-          onClick={() => router.push(`/quest/${featuredQuest?.id}`)}
-          imgSrc={featuredQuest?.img_card}
-          issuer={{
-            name: featuredQuest?.issuer ?? "",
-            logoFavicon: featuredQuest?.logo ?? "",
-          }}
-          reward={featuredQuest?.rewards_title}
-          desc={featuredQuest?.desc}
-          expiry={featuredQuest?.expiry_timestamp}
-        />
+        <div className={styles.featured_quest_banner_container}>
+          <FeaturedQuest
+            heading="Featured"
+            key={featuredQuest?.id}
+            title={featuredQuest?.title_card}
+            onClick={() => router.push(`/quest/${featuredQuest?.id}`)}
+            imgSrc={featuredQuest?.img_card}
+            issuer={{
+              name: featuredQuest?.issuer ?? "",
+              logoFavicon: featuredQuest?.logo ?? "",
+            }}
+            reward={featuredQuest?.rewards_title}
+            desc={featuredQuest?.desc}
+            expiry={featuredQuest?.expiry_timestamp}
+          />
+        </div>
+
         <QuestCategories categories={categories} />
         <div className={styles.blur2}>
           <Blur green />
@@ -43,6 +48,4 @@ const Quests: NextPage = () => {
       </div>
     </div>
   );
-};
-
-export default Quests;
+}
