@@ -2,7 +2,9 @@ import React, { FunctionComponent, useMemo } from "react";
 import { useContext } from "react";
 import { QuestsContext } from "../../context/QuestsProvider";
 import CheckIcon from "../UI/iconsComponents/icons/checkIcon";
+import UnavailableIcon from "../UI/iconsComponents/icons/unavailableIcon";
 import Card from "../UI/card";
+import styles from "../../styles/quests.module.css";
 
 type QuestProps = {
   onClick: () => void;
@@ -11,6 +13,7 @@ type QuestProps = {
   issuer: Issuer;
   reward: string;
   id: number;
+  expired: boolean;
 };
 
 const Quest: FunctionComponent<QuestProps> = ({
@@ -20,6 +23,7 @@ const Quest: FunctionComponent<QuestProps> = ({
   issuer,
   reward,
   id,
+  expired,
 }) => {
   const { completedQuestIds } = useContext(QuestsContext);
   const isCompleted = useMemo(
@@ -28,15 +32,29 @@ const Quest: FunctionComponent<QuestProps> = ({
   );
 
   return (
-    <Card imgSrc={imgSrc} title={title} onClick={onClick}>
-      <div className="flex mt-2 mb-1 items-center">
+    <Card
+      imgSrc={imgSrc}
+      title={title}
+      onClick={() => !expired && onClick()}
+      disabled={expired}
+    >
+      <div
+        className={`flex mt-2 mb-1 items-center ${
+          expired ? "opacity-40" : null
+        }`}
+      >
         <p className="text-gray-400">{issuer.name}</p>
       </div>
-      <div className="flex mt-2 mb-1 items-center">
+      <div className={styles.issuer}>
         {isCompleted ? (
           <>
             <p className="text-white mr-2">Done</p>
             <CheckIcon width="24" color="#6AFFAF" />
+          </>
+        ) : expired ? (
+          <>
+            <p className="text-white mr-2">Expired</p>
+            <UnavailableIcon width="24" color="#D32F2F" />
           </>
         ) : (
           <>
